move map_collect_result_unit to its own module
